+++ CHANGELOG.md
@@ -5,3 +5,4 @@ BACKWARDS INCOMPATIBILITIES / NOTES:
 FEATURES:
 * **New Resource**: `hopsworksai_cluster`
 * **New Data Source**: `hopsworksai_clusters`
+* **New Data Source**: `hopsworksai_aws_instance_profile_policy`